+++ R/add_difference.tbl_svysummary.R
@@ -1,23 +1,3 @@
-#' Add differences
-#'
-#' - [`add_difference.tbl_summary()`]
-#' - [`add_difference.tbl_svysummary()`]
-#'
-#' @param x (`gtsummary`)\cr
-#'   Object with class 'gtsummary'
-#' @param ... Passed to other methods.
-#' @keywords internal
-#' @author Daniel D. Sjoberg
-#' @export
-#'
-#' @seealso [`add_difference.tbl_summary()`]
-add_difference <- function(x, ...) {
-  check_not_missing(x)
-  check_class(x, "gtsummary")
-  UseMethod("add_difference")
-}
-
-
 #' Add differences between groups
 #'
 #' Adds difference to tables created by [`tbl_summary()`].
@@ -45,49 +25,20 @@ add_difference <- function(x, ...) {
 #' @return a gtsummary table of class `"tbl_summary"`
 #'
 #' @examplesIf gtsummary:::is_pkg_installed("cardx", reference_pkg = "gtsummary") && gtsummary:::is_pkg_installed("broom", reference_pkg = "cardx")
-#' # Example 1 ----------------------------------
-#' add_difference_ex1 <-
-#'   trial |>
-#'   select(trt, age, marker, response, death) %>%
-#'   tbl_summary(
-#'     by = trt,
-#'     statistic =
-#'       list(
-#'         all_continuous() ~ "{mean} ({sd})",
-#'         all_dichotomous() ~ "{p}%"
-#'       ),
-#'     missing = "no"
-#'   ) |>
-#'   # add_n() |>
-#'   add_difference()
-#'
-#' # Example 2 ----------------------------------
-#' # ANCOVA adjusted for grade and stage
-#' add_difference_ex2 <-
-#'   trial |>
-#'   select(trt, age, marker, grade, stage) %>%
-#'   tbl_summary(
-#'     by = trt,
-#'     statistic = list(all_continuous() ~ "{mean} ({sd})"),
-#'     missing = "no",
-#'     include = c(age, marker, trt)
-#'   ) |>
-#'   # add_n() |>
-#'   add_difference(adj.vars = c(grade, stage))
-add_difference.tbl_summary <- function(x,
-                                       test = NULL,
-                                       group = NULL,
-                                       adj.vars = NULL,
-                                       test.args = NULL,
-                                       conf.level = 0.95,
-                                       include = everything(),
-                                       pvalue_fun = label_style_pvalue(digits = 1),
-                                       estimate_fun = list(
-                                         c(all_continuous(), all_categorical(FALSE)) ~ label_style_sigfig(),
-                                         all_dichotomous() ~ function(x) ifelse(is.na(x), NA_character_, paste0(style_sigfig(x, scale = 100), "%")),
-                                         all_tests("smd") ~ label_style_sigfig()
-                                       ),
-                                       ...) {
+add_difference.tbl_svysummary <- function(x,
+                                          test = NULL,
+                                          group = NULL,
+                                          adj.vars = NULL,
+                                          test.args = NULL,
+                                          conf.level = 0.95,
+                                          include = everything(),
+                                          pvalue_fun = label_style_pvalue(digits = 1),
+                                          estimate_fun = list(
+                                            c(all_continuous(), all_categorical(FALSE)) ~ label_style_sigfig(),
+                                            all_dichotomous() ~ function(x) ifelse(is.na(x), NA_character_, paste0(style_sigfig(x, scale = 100), "%")),
+                                            all_tests("smd") ~ label_style_sigfig()
+                                          ),
+                                          ...) {
   set_cli_abort_call()
   # check/process inputs -------------------------------------------------------
   check_dots_empty()
@@ -101,7 +52,7 @@ add_difference.tbl_summary <- function(x,
   }
 
   # checking that input x has a by var and it has two levels
-  if (is_empty(x$inputs$by) || dplyr::n_distinct(x$inputs$data[[x$inputs$by]], na.rm = TRUE) != 2L) {
+  if (is_empty(x$inputs$by) || dplyr::n_distinct(as.data.frame(x$inputs$data)[[x$inputs$by]], na.rm = TRUE) != 2L) {
     "Cannot run {.fun add_difference} when {.code tbl_summary(by)} column does not have exactly two levels." |>
       cli::cli_abort(call = get_cli_abort_call())
   }
@@ -109,6 +60,7 @@ add_difference.tbl_summary <- function(x,
   # if `pvalue_fun` not modified, check if we need to use a theme p-value
   if (missing(pvalue_fun)) {
     pvalue_fun <-
+      get_deprecated_theme_element("add_p.tbl_svysummary-arg:pvalue_fun") %||%
       get_deprecated_theme_element("add_p.tbl_summary-arg:pvalue_fun") %||%
       get_theme_element("pkgwide-fn:pvalue_fun") %||%
       pvalue_fun
@@ -116,13 +68,13 @@ add_difference.tbl_summary <- function(x,
   pvalue_fun <- as_function(pvalue_fun)
 
   cards::process_selectors(
-    select_prep(x$table_body, x$inputs$data[x$inputs$include]),
+    select_prep(x$table_body, as.data.frame(x$inputs$data)[x$inputs$include]),
     include = {{ include }}
   )
 
   # checking for `tbl_summary(percent = c("cell", "row"))`, which don't apply
   if (!x$inputs$percent %in% "column" &&
-    any(unlist(x$inputs$type[include]) %in% c("categorical", "dichotomous"))) {
+      any(unlist(x$inputs$type[include]) %in% c("categorical", "dichotomous"))) {
     cli::cli_warn(c(
       "The {.code add_difference()} results for categorical variables may not
        compatible with {.code tbl_summary(percent = c('cell', 'row'))}.",
@@ -130,11 +82,11 @@ add_difference.tbl_summary <- function(x,
     ))
   }
 
-  cards::process_selectors(x$inputs$data, group = {{ group }}, adj.vars = {{ adj.vars }})
+  cards::process_selectors(as.data.frame(x$inputs$data), group = {{ group }}, adj.vars = {{ adj.vars }})
   check_scalar(group, allow_empty = TRUE)
 
   cards::process_formula_selectors(
-    select_prep(x$table_body, x$inputs$data[include]),
+    select_prep(x$table_body, as.data.frame(x$inputs$data)[include]),
     test = test,
     include_env = TRUE
   )
@@ -182,24 +134,24 @@ add_difference.tbl_summary <- function(x,
 
   # now process the `test.args` and `estimate_fun` arguments -------------------
   cards::process_formula_selectors(
-    select_prep(x$table_body, x$inputs$data[include]),
+    select_prep(x$table_body, as.data.frame(x$inputs$data)[include]),
     estimate_fun = estimate_fun
   )
   # fill in unspecified variables
   cards::fill_formula_selectors(
-    select_prep(x$table_body, x$inputs$data[include]),
-    estimate_fun = eval(formals(asNamespace("gtsummary")[["add_difference.tbl_summary"]])[["estimate_fun"]])
+    select_prep(x$table_body, as.data.frame(x$inputs$data)[include]),
+    estimate_fun = eval(formals(asNamespace("gtsummary")[["add_difference.tbl_svysummary"]])[["estimate_fun"]])
   )
 
   cards::process_formula_selectors(
-    select_prep(x$table_body, x$inputs$data[include]),
+    select_prep(x$table_body, as.data.frame(x$inputs$data)[include]),
     test.args = test.args
   )
   cards::check_list_elements(
     test.args,
     predicate = \(x) is.list(x) && is_named(x),
     error_msg = c("Error in the argument {.arg {arg_name}} for variable {.val {variable}}.",
-      i = "Value must be a named list."
+                  i = "Value must be a named list."
     )
   )