+++ R/assign_tests.R
@@ -105,6 +105,58 @@ assign_tests.tbl_summary <- function(x,
 
 #' @rdname assign_tests
 #' @export
+assign_tests.tbl_svysummary <- function(x,
+                                        include,
+                                        by = x$inputs$by,
+                                        test = NULL,
+                                        group = NULL,
+                                        adj.vars = NULL,
+                                        summary_type = x$inputs$type,
+                                        calling_fun = c("add_p", "add_difference"), ...) {
+  set_cli_abort_call()
+  # processing inputs ----------------------------------------------------------
+  calling_fun <- arg_match(calling_fun)
+  data <- x$inputs$data
+
+  # all variables should already have a test assigned. This looks up the tests and converts to the function
+  lapply(
+    include,
+    function(variable) {
+      if (is.null(test[[variable]])) {
+        test[[variable]] <-
+          switch(calling_fun,
+                 "add_difference" =
+                   .add_difference_tbl_svysummary_default_test(data,
+                                                               variable = variable,
+                                                               by = by, group = group, adj.vars = adj.vars,
+                                                               summary_type = summary_type[[variable]]
+                   )
+          )
+      }
+
+      if (is.null(test[[variable]])) {
+        cli::cli_abort(
+          c(
+            "There is no default test set for column {.val {variable}}.",
+            i = "Set a value in the {.arg test} argument for column {.val {variable}} or exclude with {.code include = -{variable}}."
+          ),
+          call = get_cli_abort_call()
+        )
+      }
+
+      test[[variable]] <-
+        .process_test_argument_value(
+          test = test[[variable]],
+          class = "tbl_svysummary",
+          calling_fun = calling_fun
+        )
+    }
+  ) |>
+    stats::setNames(include)
+}
+
+#' @rdname assign_tests
+#' @export
 assign_tests.tbl_continuous <- function(x,
                                         include,
                                         by,
@@ -313,6 +365,26 @@ identical_no_attr <- function(x, y) {
 
   if (!is_empty(group) && summary_type %in% c("continuous", "continuous2")) {
     return("ancova_lme4")
+  }
+
+  return(NULL)
+}
+
+.add_difference_tbl_svysummary_default_test <- function(data,
+                                                        variable,
+                                                        by,
+                                                        group,
+                                                        adj.vars,
+                                                        summary_type) {
+  if (is_empty(group) && is_empty(adj.vars) && summary_type %in% c("continuous", "continuous2")) {
+    return("svy.t.test")
+  }
+  if (is_empty(group) && is_empty(adj.vars) && summary_type %in% "categorical") {
+    return("smd")
+  }
+  # this works with and without adjustment variables
+  if (is_empty(group) && summary_type %in% c("dichotomous", "continuous", "continuous2")) {
+    return("emmeans")
   }
 
   return(NULL)